Minor fixes.
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@1301 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/wx-console/wxbutils.cpp
@@ -164,7 +164,7 @@ bool wxbPromptParser::Analyse(wxString str, int status) {
             choices->Add(""); /* index 0 is never used by multiple choice questions */
          }
          
-         if (choices->GetCount() != num) { /* new choice has begun */
+         if ((long)choices->GetCount() != num) { /* new choice has begun */
             delete choices;
             choices = new wxArrayString(num);
             choices->Add("", num); /* fill until this number */
